requirements.txt
python-dotenv
azure-cosmos
Flask
